Project files reorganization
