+++ src/watchers/batteryinfodialog.cpp
