+++ src/watchers/batteryinfodialog.h
