+++ src/watchers/batteryinfodialog.ui
