+++ src/watchers/batteryinfoframe.cpp
