+++ src/watchers/batteryinfoframe.h
