+++ src/watchers/batteryinfoframe.ui
